+++ 404.php
@@ -10,7 +10,7 @@
 
 							<header class="article-header">
 
-								<h1><?php _e( 'Epic 404 - Article Not Found', 'braftonium' ); ?></h1>
+								<h1><?php _e( 'Page Not Found', 'braftonium' ); ?></h1>
 
 							</header>
 
@@ -29,7 +29,26 @@
 						</article>
 
 					</main>
-
+					<div class="sidebar m-all t-1of3 d-2of7 last-col cf" role="complementary">
+						<h2><?php _e( 'Or were you looking for:', 'braftonium' ); ?></h2>
+						<h3><?php _e( 'Our Latest Post', 'braftonium' ); ?></h3>
+						<?php
+							$args = array( 'numberposts' => '1' );
+							$recent_posts = wp_get_recent_posts( $args );
+							foreach( $recent_posts as $recent ){
+								echo '<a href="' . get_permalink($recent["ID"]) . '" class="blue-btn">'. $recent["post_title"].'</a>';
+							}
+							wp_reset_query();
+						?>
+						<h3><?php _e( 'A Way to Contact Us', 'braftonium' ); ?></h3>
+							<?php $page = get_page_by_title_search('contact');
+							if ($page):
+								echo '<a href="' . get_permalink($page[0]->ID) . '" class="blue-btn">'. $page[0]->post_title.'</a>';
+							endif; ?>
+						<h3><?php _e( 'Our Homepage', 'braftonium' ); ?></h3>
+						<?php $frontpage_id = get_option( 'page_on_front' ); 
+							echo '<a href="' . get_permalink($frontpage_id) . '" class="blue-btn">'.get_the_title($frontpage_id).'</a>'; ?>
+					</div>
 				</div>
 
 			</div>