+++ library/js/sticky.js
@@ -1,18 +1,18 @@
-jQuery(document).ready(function($){
+jQuery(document).ready(function ($) {
   // if ($(window).width() > 768) {
-    var mn = $('#inner-header');
-        mns = 'scrolled';
-        hdr = $('#inner-header').height();
-        content = $('header .banner');
-  
-    $(window).scroll(function() {
-      if( $(this).scrollTop() > hdr-45 ) {
-        mn.addClass(mns);
-        content.css( "margin-top", hdr );
-      } else {
-        mn.removeClass(mns);
-        content.css( "margin-top", '0' );
-      }
-    });
+  var mn = $('#inner-header');
+  mns = 'scrolled';
+  hdr = $('#inner-header').height();
+  content = $('#container');
+
+  $(window).scroll(function () {
+    if ($(this).scrollTop() > hdr) {
+      mn.addClass(mns);
+      content.css("margin-top", hdr);
+    } else {
+      mn.removeClass(mns);
+      content.css("margin-top", '0');
+    }
+  });
   // }
 });