+++ post-formats/content-map.php
@@ -28,9 +28,6 @@ if ( $style['other'] ) {
 	if (in_array('compact', $style['other'])){
 		$classes[] = "compact";
 	}
-	if (in_array('thin', $style['other'])){
-		$classes[] = "thin";
-	}
 	if (in_array('center', $style['other'])){
 		$classes[] = "center";
 	}